Working on taking over tempo.
Much of the packet and tracking infrastructure is in place and
working; telling the virtual CDJ to take over the tempo (either by
directly calling the function, or sending it a the packet that tells
it to) generates the right message to the current master, and we get
an ack, but the change never happens.

I suspect the problem is that my shortcut of trying to send mixer
status packets rather than CDJ status packets is just not going to
work, so the next step is to put together full CDJ status packets (and
set up so we display the ones we are sending too, so I can visually
compare them with real ones for debugging), and see where that leaves
us.